Address OPE-124 likelihood ownership review
