+++ newsfragments/OPE-124.misc.md
@@ -1,1 +1,1 @@
-Make additive and pollution-event mismatch equations first-class peer model components, load minimum-error floors only for components which own them, narrow custom likelihood inputs to universal observation terms, and isolate frozen additive translation at the RHIME compatibility boundary.
+Make additive and pollution-event mismatch equations first-class peer model components, load minimum-error floors only for components which own them, narrow custom likelihood inputs to universal observation terms, and translate legacy ``run_hbmcmc`` additive configuration directly into explicit model options.
