+++ openghg_inversions/rhime/__init__.py
@@ -24,7 +24,7 @@ from .builders import (
     RhimeModelBuilderContext,
     RhimeModelBuildResult,
 )
-from .likelihood_seam import RhimeLikelihoodBuilder
+from ._custom_likelihood_seam import RhimeLikelihoodBuilder
 from .co2 import build_co2_model, co2_model_input_names, run_rhime_co2
 from .materialization import materialize_pymc_inputs
 from .params import params_from_config, resolve_flux_sources, resolve_rhime_options
