--- openghg_inversions/rhime/likelihoods.py
@@ -1,6 +1,0 @@
-"""Frozen compatibility import for the unchanged ``run_hbmcmc`` entry point."""
-
-from openghg_inversions.rhime.likelihood_seam import additive_sigma_likelihood_builder
-
-
-__all__ = ["additive_sigma_likelihood_builder"]
